requirements.txt
alembic==1.16.1
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
